+++ lib/core/enums/turbo_flag_type.dart
@@ -6,8 +6,7 @@ enum TurboFlagType {
   clean,
   force;
 
-  static List<TurboFlagType> get globalValues =>
-      values.where((flag) => flag.isGlobal).toList();
+  static List<TurboFlagType> get globalValues => values.where((flag) => flag.isGlobal).toList();
 
   List<String> bashCommands({required TurboCommandType source}) {
     switch (source) {
@@ -18,6 +17,7 @@ enum TurboFlagType {
         ];
       case TurboCommandType.clone:
       case TurboCommandType.archive:
+      case TurboCommandType.update:
         return [];
     }
   }